envelope struct and storage - use double for sample rate

+++ src/synth/src/instrument.c
@@ -7,11 +7,11 @@
 #define TICKS_PER_PART 2
 #define VOLUME_LOG_SCALE 0.1428
 
-double calc_samples_per_tick(double bpm, int sample_rate) {
+double calc_samples_per_tick(double bpm, double sample_rate) {
     const double beats_per_sec = bpm / 60.0;
     const double parts_per_sec = PARTS_PER_BEAT * beats_per_sec;
     const double ticks_per_sec = TICKS_PER_PART * parts_per_sec;
-    return (double)sample_rate / ticks_per_sec;
+    return sample_rate / ticks_per_sec;
 }
 
 double note_size_to_volume_mult(double size) {
